+++ lega/an/fade_in.py
@@ -1,7 +1,4 @@
 from lega.an.animation import Animation
-from lega.core import RuntimeUnit
-
-from lega.globe import sttmgr
 
 from config import color_theme
 
@@ -9,13 +6,16 @@ from typing import Union, Tuple
 
 from pygame import Surface, Rect
 
+from lega.globe import scrmgr
+from lega.event_manager import handle_events
+
 
 class FadeIn(Animation):
     """
     实现内容从屏幕上“淡出”的效果。
     """
     def __init__(
-            self, rtu: RuntimeUnit,
+            self,
             variable_surf: Surface,
             destination: Union[Rect, Tuple[int, int]] = (0, 0),
             speed: int = 20,
@@ -27,21 +27,24 @@ class FadeIn(Animation):
         :param speed:
         :return:
         """
-        Animation.__init__(self, rtu)
+        Animation.__init__(self)
 
         self.variable_surf = variable_surf
         self.destination = destination
         self.speed = speed
 
     def play(self):
+        surf = self.variable_surf.copy()
         for alpha in range(0, 255, self.speed):
-            self.rtu.screen.fill(color_theme.background)
-            self.variable_surf.set_alpha(alpha)
-            self.rtu.screen.blit(self.variable_surf, self.destination)
-            self.rtu.flip()
+            scrmgr.screen.fill(color_theme.background)
+            surf.set_alpha(alpha)
+            scrmgr.screen.blit(surf, self.destination)
+            scrmgr.update_global()
 
-            self.rtu.handle_universal_events_during_each_animation_frame()
-
-            if sttmgr.should_return_at_once:
-                sttmgr.should_return_at_once = False  # 还原状态
+            """
+            # FadeIn 似乎不需要下面这部分逻辑
+            handle_events(*self._handlers)
+            if self.should_return_at_once:
+                self.should_return_at_once = False  # 还原状态
                 return
+            """